requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
python-multipart==0.0.6
openai==1.12.0
python-dotenv==1.0.0
jinja2==3.1.3
aiofiles==23.2.1
stripe==8.0.0
